Moved to proper place for future refactoring
