+++ src/check_entry_exit.cpp
@@ -1,5 +1,5 @@
 #include "check_entry_exit.h"
-#include "exchange.h"
+#include "exchanges/exchange.h"
 #include "result.h"
 #include "parameters.h"
 #include <sstream>
