+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "exchange.h"
+#include "exchanges/exchange.h"
 #include "result.h"
 #include "time_fun.h"
 #include "curl_fun.h"
